support rpm 4.11
We were relying on the newer rpm.files Python API which is present in
rpm 4.12+, but RHEL7 has rpm 4.11. It's desirable to have the conflict
detection code working with the older rpm APIs.

To call rpmfiCompare() from Python we have to use some nasty ctypes
hacking so we still prefer the original rpm.files code path whenever
that is available.

Bug: 1369317
Change-Id: I3927c2a77267d2f9592055f6aac08fe0c6818010