allow rpmdeplint to detect arch of rpms under test
Previously if you ran repoclosure on a different
architecture it would incorrectly warn about
pre-existing repoclosure problems.
This was due to the architecture being defaultly
set to the host machines architecture within hawkey.
Which made queries with provides not return the
expected results.

Added in the arch parameter to allow users to
set the arch if required.

Bug: 1378253
Change-Id: I1da559a925863845965210eab613eb2b1c61c7ff