packaging: supply separate Python 2 and Python 3 subpackages
On Fedora, the CLI will still be provided by rpmdeplint but it will
depend on python3-rpmdeplint which contains the Python modules.
python2-rpmdeplint will also be available for other scripts which want
to use the Python modules in Python 2.

On EPEL7, only python2-rpmdeplint will be built, and the base package
will use that instead.

Change-Id: Id943fe9a48bc7d748ed266eb54bf5d12bbdc3aa5