noarch is not a real arch we can test against
If we are given only noarch packages, we need the caller to tell us
which arch to test against because we can't guess it.

If we are given a combination of archful packages and noarch packages,
that should not be an error because noarch is by definition compatible
with everything.

Bug: 1392635
Change-Id: I0a426a23b7e8674a54aa33a4cfb7cea375f665da