tests: use rpmfluff instead of hawkey.test.TestSackMixin
The fake repo format used in hawkey's tests is really an internal detail
(actually internal to libsolv) which we shouldn't be relying on in the
unit tests.

We can use rpmfluff instead, same as we do in the acceptance tests.

This patch depends on a fix in rpmfluff to handle noarch packages in
repos: https://pagure.io/rpmfluff/pull-request/15

Change-Id: I2e8d7b21703a8cc4abbdb881e1f32ddec38d9bc6