remove --arch guesswork
Previously we had to tell the hawkey.Sack to solve dependencies for
*some* specific arch, which is why we added the --arch option. We were
able to guess the desired arch in some cases, but it left us with some
awkward edge cases which we couldn't support, like testing noarch
packages without an explicit arch or testing multiple arches at the same
time.

Now that we have ported to libsolv, the libsolv.Pool will happily accept
.setarch(None) which means that dependency resolution will pick any
architecture available from the repos. Ideally this is what we would
have done from the beginning.

We keep the --arch option for compatibility, and it will continue to
behave the same way (namely, when an arch is given, dependency
resolution will only consider packages compatible with that arch and
ignore all others). But now --arch is optional, we don't have to do any
guessing, and those awkward edge cases are now irrelevant.

Change-Id: I64511babbad58a575e0546a16537fc0d671e319c