report a proper error if we fail to read a package
It's not well documented, but in the libsolv Python bindings .add_rpm()
will return None instead of a Solvable if there is an error reading the
package.

This is actually exactly the same bug which affected hawkey and
originally caused the crash in rpmdeplint, but of course we are in
Python land now so instead of a crash we just got an unexpected None
value.

Explicitly check for a None return and raise a proper informative error
in that case.

Bug: 1382531
Change-Id: Id7d8b83add5c2b8b4f14e6bb2417805eb6e0b722