handle "installonly" packages properly
This is how DNF indicates to libsolv that kernel and related packages
are "install only", that is, can be installed in parallel. We need to
apply the same handling so that we correctly check conflicts between two
kernel packages, rather than ignoring the conflict assuming they can't
be installed together.

Bug: 1465734
Change-Id: I5feaf8eef0abacd087ba77484d5f91a6051052ff