remove extra layer of directories in cache
Instead of:

    ~/.cache/rpmdeplint/f/682d3295c6dad6dcaa87fe4c38ab5a295b9295e35e70d566f11ba51c72e606b/f682d3295c6dad6dcaa87fe4c38ab5a295b9295e35e70d566f11ba51c72e606b-primary.xml.gz

it becomes:

    ~/.cache/rpmdeplint/f/682d3295c6dad6dcaa87fe4c38ab5a295b9295e35e70d566f11ba51c72e606b

Also adjusted the cache to pass back an open file object instead of
a filename. This closes a race window between opening the cache entry
(to check if it exists) and hawkey re-opening it to read the contents.

Previously we couldn't fix this because hawkey requires a filename and
expects to open it itself. But with libsolv we can use solv.xfopen_fd()
to pass an open file descriptor instead.

See https://bugzilla.redhat.com/show_bug.cgi?id=1343247#c12

Change-Id: I40f2e0d226b6d18b86fb389ac9e52104073716f1