rearrange conflict checking algorithm
Instead of iterating each file in the package under test, and finding
every package containing that file (which is a relatively slow operation
in libsolv) -- we can instead iterate *every* other package and use
Python set operations to look for any overlaps in filenames.

This substantially improves the running time of check-conflicts when
testing packages with a large number of files.

Bug: 1465736
Change-Id: I49634b18738d997f7b8596fcebd7abd09da93385