accept unrecognised arches as-is
The purpose of the arch mapping is just to map the various compatible
sub-architectures to their main architecture. However there are many
arches which have *no* sub-architectures that are currently omitted from
the list, causing an error. In those cases, we should just use whatever
we find on the RPM package as is instead.

See:
https://github.com/openSUSE/libsolv/blob/aea953a42d99399938eb6c98a220310aa438c376/src/poolarch.c#L23
https://github.com/openSUSE/libsolv/commit/aea953a42d99399938eb6c98a220310aa438c376

Bug: 1562073
Change-Id: Id094f0d8bfde977cbedb5408d11688ccd3c048c4